pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-qahelper"
version = "0.1.0"
description = "Локальный CLI для генерации тест-кейсов из требований и макетов (OpenAI, опционально Figma API)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "q1nn2" }]
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "pydantic>=2",
    "httpx>=0.27",
    "openai>=1.40",
    "pypdf>=4",
    "pyyaml>=6",
    "pandas>=2",
    "openpyxl>=3",
    "gspread>=6",
    "google-auth>=2",
]

[project.optional-dependencies]
autotest = ["pytest>=8", "pytest-html>=4", "playwright>=1.40"]

[project.scripts]
ai-qahelper = "ai_qahelper.cli:run_cli"

[tool.setuptools.packages.find]
where = ["src"]
